refactor to args
